refactor: reworks csv_to_yaml for pydantic
Update write empty keys to write default keys
Create transformer for CSV reading and writing
Moves CSV specific logic t csv_transformer
Moves YAML specific logic to yaml_transformer and adds writing logic
