+++ tests/data/yaml/test_complete_rule.yaml
@@ -4,12 +4,12 @@ x-trestle-rule-info:
   parameter:
     name: prm_1
     description: prm_1 description
-    alternative-values: {'default': '5%', '5pc': '5%', '10pc': '10%', '15pc': '15%', '20pc': '20%'}
-    default-value: '5%'
+    alternative_values: {'default': '5%', '5pc': '5%', '10pc': '10%', '15pc': '15%', '20pc': '20%'}
+    default_value: '5%'
   profile:
     description: Simple NIST Profile
     href: profiles/simplified_nist_profile/profile.json
-    include-controls:
+    include_controls:
       - id: ac-1
 x-trestle-component-info:
   name: Component 1
