+++ tests/trestlebot/transformers/test_csv_to_yaml.py
@@ -16,13 +16,12 @@
 
 import csv
 import pathlib
-from dataclasses import fields
 
 import pytest
 import ruamel.yaml as yaml
 
 from trestlebot.transformers.csv_to_yaml import YAMLBuilder
-from trestlebot.transformers.trestle_rule import TrestleRule
+from trestlebot.transformers.yaml_transformer import RulesYAMLTransformer
 
 
 @pytest.fixture(scope="function")
@@ -76,13 +75,19 @@ def test_write_to_yaml(setup_yaml_builder: YAMLBuilder, tmp_trestle_dir: str) ->
     assert len(data) == 1
 
 
-def test_write_empty_trestle_rule_keys(
+def test_default_test_trestle_rule_keys(
     setup_yaml_builder: YAMLBuilder, tmp_trestle_dir: str
 ) -> None:
     yaml_file = pathlib.Path(tmp_trestle_dir) / "test.yaml"
-    setup_yaml_builder.write_empty_trestle_rule_keys(yaml_file)
-    with open(yaml_file, "r") as f:
-        data = yaml.safe_load(f)
-    assert all(value == "" for value in data.values())
-    expected_keys = {field.name for field in fields(TrestleRule)}
-    assert expected_keys == set(data.keys())
+    setup_yaml_builder.write_default_trestle_rule_keys(yaml_file)
+    transformer = RulesYAMLTransformer()
+    rule = transformer.transform_to_rule(yaml_file.read_text())
+
+    assert rule.name == "example rule"
+    assert rule.description == "example description"
+    assert rule.component.name == "example component"
+    assert rule.component.description == "example description"
+    assert rule.component.type == "service"
+    assert rule.profile.description == "example profile"
+    assert rule.profile.href == "example href"
+    assert len(rule.profile.include_controls) == 1
