+++ trestlebot/const.py
@@ -34,8 +34,8 @@ LEVERAGED_SSP_KEY_NAME = "leveraged_ssp"
 RULE_INFO_TAG = trestle_const.TRESTLE_TAG + "rule-info"
 NAME = "name"
 DESCRIPTION = "description"
-PARAMETERS = "parameter"
-PROFILES = "profile"
+PARAMETER = "parameter"
+PROFILE = "profile"
 HREF = "href"
 INCLUDE_CONTROLS = "include-controls"
 DEFAULT_VALUE = "default-value"
