+++ trestlebot/transformers/base_transformer.py
@@ -1,0 +1,38 @@
+#!/usr/bin/python
+
+#    Copyright 2023 Red Hat, Inc.
+#
+#    Licensed under the Apache License, Version 2.0 (the "License"); you may
+#    not use this file except in compliance with the License. You may obtain
+#    a copy of the License at
+#
+#         http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+#    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+#    License for the specific language governing permissions and limitations
+#    under the License.
+
+"""Base transformer for rules."""
+
+from abc import ABC, abstractmethod
+from typing import Any
+
+from trestlebot.transformers.trestle_rule import TrestleRule
+
+
+class RulesTransformer(ABC):
+    """Abstract interface for transformers for rules"""
+
+    @abstractmethod
+    def transform_to_rule(self, data: Any) -> TrestleRule:
+        """Transform to rule data."""
+
+    @abstractmethod
+    def transform_from_rule(self, rule: TrestleRule) -> Any:
+        """Transform from rule data."""
+
+
+class RulesTransformerException(Exception):
+    """An error during transformation of a rule"""
